Implemented specified location/load adjustments
